Refactor the duplicate services code into a mixin
