+++ test/tests/unit/services/popup-test.js
@@ -1,6 +1,6 @@
 import Popup from 'torii/services/popup';
 import PopupIdSerializer from 'torii/lib/popup-id-serializer';
-import { CURRENT_REQUEST_KEY } from 'torii/services/popup';
+import { CURRENT_REQUEST_KEY } from 'torii/mixins/services-mixin';
 
 var popup;
 var originalWindowOpen = window.open;
@@ -50,7 +50,7 @@ asyncTest("open resolves based on popup window", function(){
   var popupId = '09123-asdf';
   var mockWindow = null
 
-  popup = new Popup({popupIdGenerator: buildPopupIdGenerator(popupId)});
+  popup = new Popup({remoteIdGenerator: buildPopupIdGenerator(popupId)});
 
   window.open = function(url, name){
     ok(true, 'calls window.open');
